requirements.txt
streamlit==1.44.0
openai==1.68.2
python-dotenv==1.1.0